[new] range和size都支持固定大小或者区间段  xxx=value 或 xxx=[a,b] 的写法
