+++ src/Luban.Job.Cfg/Source/Validators/RangeValidator.cs
@@ -11,128 +11,17 @@ namespace Luban.Job.Cfg.Validators
     {
         public TType Type { get; }
 
-        private readonly string _str;
-
-        private long? _min;
-        private long? _max;
-
-        private double? _mind;
-        private double? _maxd;
-
-        private bool _includeMinBound;
-
-        private bool _includeMaxBound;
+        private Range _range;
 
         public RangeValidator(TType type, string strRange)
         {
             Type = type;
-            _str = strRange.Trim();
-        }
-
-        private bool TryParse(string s, ref long? x)
-        {
-            s = s.Trim();
-            if (string.IsNullOrEmpty(s))
-            {
-                x = null;
-                return true;
-            }
-            else if (long.TryParse(s, out var v))
-            {
-                x = v;
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        private bool TryParse(string s, ref double? x)
-        {
-            s = s.Trim();
-            if (string.IsNullOrEmpty(s))
-            {
-                x = null;
-                return true;
-            }
-            else if (double.TryParse(s, out var v))
-            {
-                x = v;
-                return true;
-            }
-            else
-            {
-                return false;
-            }
+            _range = new Range(strRange);
         }
 
         public void Compile(DefFieldBase def)
         {
-            void ThrowError()
-            {
-                throw new Exception($"结构:{ def.HostType.FullName } 字段: { def.Name}  range 定义:{_str} 不合法");
-            }
-
-            if (_str.Length <= 2)
-            {
-                ThrowError();
-            }
-            switch (_str[0])
-            {
-                case '[': _includeMinBound = true; break;
-                case '(': _includeMinBound = false; break;
-                default: ThrowError(); break;
-            }
-            switch (_str[^1])
-            {
-                case ']': _includeMaxBound = true; break;
-                case ')': _includeMaxBound = false; break;
-                default: ThrowError(); break;
-            }
-
-            var pars = _str[1..^1].Split(',');
-            if (pars.Length != 2)
-            {
-                ThrowError();
-            }
-
-            bool p1 = TryParse(pars[0], ref _min);
-            bool p2 = TryParse(pars[0], ref _mind);
-            bool p3 = TryParse(pars[1], ref _max);
-            bool p4 = TryParse(pars[1], ref _maxd);
-
-            if ((!p1 && !p2) || (!p3 && !p4))
-            {
-                ThrowError();
-            }
-        }
-
-
-        private bool CheckInLongRange(long x)
-        {
-            if (_min is long m && (_includeMinBound ? m > x : m >= x))
-            {
-                return false;
-            }
-            if (_max is long n && (_includeMaxBound ? n < x : n <= x))
-            {
-                return false;
-            }
-            return true;
-        }
-
-        private bool CheckInDoubleRange(double x)
-        {
-            if (_mind is double m && (_includeMinBound ? m > x : m >= x))
-            {
-                return false;
-            }
-            if (_maxd is double n && (_includeMaxBound ? n < x : n <= x))
-            {
-                return false;
-            }
-            return true;
+            _range.Compile();
         }
 
         public string Source => ValidatorContext.CurrentVisitor.CurrentValidateRecord.Source;
@@ -142,7 +31,7 @@ namespace Luban.Job.Cfg.Validators
             var assembly = ctx.Assembly;
             void LogError()
             {
-                assembly.Agent.Error("记录 {0}:{1} (来自文件:{2}) 不在范围:{3}内", ValidatorContext.CurrentRecordPath, data, Source, _str);
+                assembly.Agent.Error("记录 {0}:{1} (来自文件:{2}) 不在范围:{3}内", ValidatorContext.CurrentRecordPath, data, Source, _range.RawStr);
             }
 
             if (type.IsNullable && data == null)
@@ -154,7 +43,7 @@ namespace Luban.Job.Cfg.Validators
             {
                 case DByte b:
                 {
-                    if (!CheckInLongRange(b.Value))
+                    if (!_range.CheckInLongRange(b.Value))
                     {
                         LogError();
                         return;
@@ -163,7 +52,7 @@ namespace Luban.Job.Cfg.Validators
                 }
                 case DFshort s:
                 {
-                    if (!CheckInLongRange(s.Value))
+                    if (!_range.CheckInLongRange(s.Value))
                     {
                         LogError();
                         return;
@@ -172,7 +61,7 @@ namespace Luban.Job.Cfg.Validators
                 }
                 case DShort s:
                 {
-                    if (!CheckInLongRange(s.Value))
+                    if (!_range.CheckInLongRange(s.Value))
                     {
                         LogError();
                         return;
@@ -181,7 +70,7 @@ namespace Luban.Job.Cfg.Validators
                 }
                 case DInt i:
                 {
-                    if (!CheckInLongRange(i.Value))
+                    if (!_range.CheckInLongRange(i.Value))
                     {
                         LogError();
                         return;
@@ -190,7 +79,7 @@ namespace Luban.Job.Cfg.Validators
                 }
                 case DFint i:
                 {
-                    if (!CheckInLongRange(i.Value))
+                    if (!_range.CheckInLongRange(i.Value))
                     {
                         LogError();
                         return;
@@ -199,7 +88,7 @@ namespace Luban.Job.Cfg.Validators
                 }
                 case DLong l:
                 {
-                    if (!CheckInLongRange(l.Value))
+                    if (!_range.CheckInLongRange(l.Value))
                     {
                         LogError();
                         return;
@@ -208,7 +97,7 @@ namespace Luban.Job.Cfg.Validators
                 }
                 case DFlong fl:
                 {
-                    if (!CheckInLongRange(fl.Value))
+                    if (!_range.CheckInLongRange(fl.Value))
                     {
                         LogError();
                         return;
@@ -217,7 +106,7 @@ namespace Luban.Job.Cfg.Validators
                 }
                 case DFloat ff:
                 {
-                    if (!CheckInDoubleRange(ff.Value))
+                    if (!_range.CheckInDoubleRange(ff.Value))
                     {
                         LogError();
                         return;
@@ -226,7 +115,7 @@ namespace Luban.Job.Cfg.Validators
                 }
                 case DDouble dd:
                 {
-                    if (!CheckInDoubleRange(dd.Value))
+                    if (!_range.CheckInDoubleRange(dd.Value))
                     {
                         LogError();
                         return;
